requirements.txt
aiohttp==3.6.2
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pylint==2.5.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
